:truck: move AnswerCollectionForScoring to specific contexte
