+++ api/src/certification/shared/domain/models/AnswerCollectionForScoring.js
