+++ api/tests/unit/domain/models/AnswerCollectionForScoring_test.js
@@ -1,4 +1,4 @@
-import { AnswerCollectionForScoring } from '../../../../src/shared/domain/models/AnswerCollectionForScoring.js';
+import { AnswerCollectionForScoring } from '../../../../src/certification/shared/domain/models/AnswerCollectionForScoring.js';
 import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
 import { domainBuilder, expect } from '../../../test-helper.js';
